+++ __test__/users.test.ts
@@ -146,52 +146,3 @@ describe("[GET] /api/users/[id]", () => {
     );
   });
 });
-
-describe("[GET] /api/staff", () => {
-  it("look for all staff", async () => {
-    const expected: User[] = [
-      {
-        id: "2",
-        first_name: "Teacher",
-        last_name: "Doe",
-        email: "teacher@gmail.com",
-        role: "Teacher",
-        address: "123 Main Street",
-        phone_number: "1234567890",
-      },
-      {
-        id: "3",
-        first_name: "Admin",
-        last_name: "Doe",
-        email: "admin@gmail.com",
-        role: "Admin",
-        address: "123 Main Street",
-        phone_number: "1234567890",
-      },
-      {
-        id: "50",
-        first_name: "Admin",
-        last_name: "Doe",
-        email: "newAdmin@gmail.com",
-        role: "Admin",
-        address: "123 Main Street",
-        phone_number: "1234567890",
-      },
-      {
-        id: "45",
-        first_name: "Teacher",
-        last_name: "Doe",
-        email: "newTeacher@gmail.com",
-        role: "Admin",
-        address: "123 Main Street",
-        phone_number: "1234567890",
-      },
-    ];
-    await makeHTTPRequest(staffHandler, "/api/staff", undefined, "GET", undefined, StatusCodes.ACCEPTED, expected);
-  });
-
-  it("look for all staff when there are none", async () => {
-    client.query("DELETE from users");
-    await makeHTTPRequest(staffHandler, "/api/users/1", undefined, "GET", undefined, StatusCodes.ACCEPTED, []);
-  });
-});
